requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
python-jose[cryptography]==3.3.0
python-dotenv==1.1.1
